HW11 Fixed validation of input parameters
